requirements.txt
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
click==8.2.1
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-telegram-bot==20.7
sniffio==1.3.1
typing_extensions==4.14.1
Werkzeug==3.1.3
